refactor(launcher): move handleCancel from utils.ts to prompts.ts
